Renamed two handlers to helpers
