+++ src/ui/handlers/pokedex-ui-handler.ts
@@ -50,7 +50,7 @@ import type { OptionSelectConfig } from "#ui/handlers/abstract-option-select-ui-
 import { MessageUiHandler } from "#ui/handlers/message-ui-handler";
 import { addTextObject, getTextColor } from "#ui/text";
 import { addWindow } from "#ui/ui-theme";
-import { PokemonIconAnimHandler, PokemonIconAnimMode } from "#ui/utils/pokemon-icon-anim-handler";
+import { PokemonIconAnimHelper, PokemonIconAnimMode } from "#ui/utils/pokemon-icon-anim-helper";
 import { BooleanHolder, fixedInt, getLocalizedSpriteKey, padInt, randIntRange, rgbHexToRgba } from "#utils/common";
 import type { StarterPreferences } from "#utils/data";
 import { loadStarterPreferences } from "#utils/data";
@@ -198,7 +198,7 @@ export class PokedexUiHandler extends MessageUiHandler {
   public cursorObj: Phaser.GameObjects.Image;
   private pokerusCursorObjs: Phaser.GameObjects.Image[];
 
-  private iconAnimHandler: PokemonIconAnimHandler;
+  private iconAnimHandler: PokemonIconAnimHelper;
 
   private starterPreferences: StarterPreferences;
 
@@ -482,7 +482,7 @@ export class PokedexUiHandler extends MessageUiHandler {
       pokemonContainerWindow.setVisible(false);
     }
 
-    this.iconAnimHandler = new PokemonIconAnimHandler();
+    this.iconAnimHandler = new PokemonIconAnimHelper();
     this.iconAnimHandler.setup();
 
     this.pokemonNumberText = addTextObject(6, 141, "", TextStyle.SUMMARY);
