+++ src/ui/handlers/starter-select-ui-handler.ts
@@ -67,7 +67,7 @@ import type { OptionSelectItem } from "#ui/handlers/abstract-option-select-ui-ha
 import { MessageUiHandler } from "#ui/handlers/message-ui-handler";
 import { addBBCodeTextObject, addTextObject, getTextColor } from "#ui/text";
 import { addWindow } from "#ui/ui-theme";
-import { PokemonIconAnimHandler, PokemonIconAnimMode } from "#ui/utils/pokemon-icon-anim-handler";
+import { PokemonIconAnimHelper, PokemonIconAnimMode } from "#ui/utils/pokemon-icon-anim-helper";
 import { applyChallenges, checkStarterValidForChallenge } from "#utils/challenge-utils";
 import {
   BooleanHolder,
@@ -398,7 +398,7 @@ export class StarterSelectUiHandler extends MessageUiHandler {
   private startCursorObj: Phaser.GameObjects.NineSlice;
   private randomCursorObj: Phaser.GameObjects.NineSlice;
 
-  private iconAnimHandler: PokemonIconAnimHandler;
+  private iconAnimHandler: PokemonIconAnimHelper;
 
   //variables to keep track of the dynamically rendered list of instruction prompts for starter select
   private instructionRowX = 0;
@@ -611,7 +611,7 @@ export class StarterSelectUiHandler extends MessageUiHandler {
       starterContainerWindow.setVisible(false);
     }
 
-    this.iconAnimHandler = new PokemonIconAnimHandler();
+    this.iconAnimHandler = new PokemonIconAnimHelper();
     this.iconAnimHandler.setup();
 
     this.pokemonSprite = globalScene.add.sprite(53, 63, "pkmn__sub");
